refactor: organize components in new folders
