+++ src/pages/Session/components/Seat.js
