+++ src/pages/Success/components/HomeButton.js
@@ -1,4 +1,4 @@
-import "../css/button.css";
+import "../../../css/button.css";
 import { Link } from "react-router-dom";
 
 export function HomeButton() {
